Renaming & move button bindings to the right method
