+++ src/main/java/frc/robot/shooter/ShootCommand.java
@@ -2,27 +2,25 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.intake;
+package frc.robot.shooter;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.constants.PhysicalConstants.IntakeConstants;
+import frc.robot.constants.PhysicalConstants.ShooterConstants;
 
-/** Spins the intake motors. */
-public class SpinIntakeCommand extends Command {
-    /**
-     * Creates a new SpinIntakeCommand.
-     */
-    public SpinIntakeCommand() {
-        setName("SpinIntakeCommand");
+/** Spins the shooter motors. */
+public class ShootCommand extends Command {
+    /** Creates a new ShootCommand. */
+    public ShootCommand() {
+        setName("ShootCommand");
         
         // Use addRequirements() here to declare subsystem dependencies.
-        addRequirements(IntakeSubsystem.getInstance());
+        addRequirements(ShooterSubsystem.getInstance());
     }
 
     // Called when the command is initially scheduled.
     @Override
     public void initialize() {
-        IntakeSubsystem.getInstance().setIntakeSpeed(IntakeConstants.INTAKE_SPEED);
+        ShooterSubsystem.getInstance().setShooterSpeed(ShooterConstants.SHOOTER_SPEED);
     }
 
     // Called every time the scheduler runs while the command is scheduled.
@@ -32,7 +30,7 @@ public class SpinIntakeCommand extends Command {
     // Called once the command ends or is interrupted.
     @Override
     public void end(boolean interrupted) {
-        IntakeSubsystem.getInstance().setIntakeSpeed(0);
+        ShooterSubsystem.getInstance().setShooterSpeed(0);
     }
 
     // Returns true when the command should end.
